MM_33223 : Refactoring file types.ts
